# tools/WillowWSUP/WsupApi/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Willow.Support.Api",
    "version": "v1"
  },
  "paths": {
...
    "/Connectors": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connector"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_Create",
        "requestBody": {
          "x-name": "connector",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Connector"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connector"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connector"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_Update",
        "requestBody": {
          "x-name": "connector",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Connector"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connector"
                }
              }
            }
          }
        }
      }
    },
    "/Connectors/GetAllByCustomerInstance/{customerInstanceId}": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_GetAllByCustomerInstance",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connector"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Connectors/{connectorId}": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_Get",
        "parameters": [
          {
            "name": "connectorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Connector"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Connectors"
        ],
        "operationId": "Connectors_Delete",
        "parameters": [
          {
            "name": "connectorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/ConnectorTypes": {
      "get": {
        "tags": [
          "ConnectorTypes"
        ],
        "operationId": "ConnectorTypes_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectorType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ConnectorTypes"
        ],
        "operationId": "ConnectorTypes_Create",
        "requestBody": {
          "x-name": "connectorType",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorType"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorType"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ConnectorTypes"
        ],
        "operationId": "ConnectorTypes_Update",
        "requestBody": {
          "x-name": "connectorType",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorType"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorType"
                }
              }
            }
          }
        }
      }
    },
    "/ConnectorTypes/{connectorTypeId}": {
      "get": {
        "tags": [
          "ConnectorTypes"
        ],
        "operationId": "ConnectorTypes_Get",
        "parameters": [
          {
            "name": "connectorTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConnectorType"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ConnectorTypes"
        ],
        "operationId": "ConnectorTypes_Delete",
        "parameters": [
          {
            "name": "connectorTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/CustomerInstanceApplications/GetAllApplicationsByCustomer/{customerInstanceId}": {
      "get": {
        "tags": [
          "CustomerInstanceApplications"
        ],
        "operationId": "CustomerInstanceApplications_GetAll2",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerInstanceApplication"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/CustomerInstanceApplications/{customerInstanceId}/{applicationId}": {
      "get": {
        "tags": [
          "CustomerInstanceApplications"
        ],
        "operationId": "CustomerInstanceApplications_Get",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInstanceApplication"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CustomerInstanceApplications"
        ],
        "operationId": "CustomerInstanceApplications_Delete",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/CustomerInstanceApplications": {
      "post": {
        "tags": [
          "CustomerInstanceApplications"
        ],
        "operationId": "CustomerInstanceApplications_Create",
        "requestBody": {
          "x-name": "customerInstanceApplication",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInstanceApplication"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "CustomerInstanceApplications"
        ],
        "operationId": "CustomerInstanceApplications_Update",
        "requestBody": {
          "x-name": "customerInstanceApplication",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInstanceApplication"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/CustomerInstances/GetAll": {
      "get": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_GetAll",
        "parameters": [
          {
            "name": "includeApplications",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerInstance"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Application": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "teamId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "healthEndpointPath": {
            "type": "string",
            "nullable": true
          },
          "hasPublicEndpoint": {
            "type": "boolean"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Team"
              }
            ]
          },
          "applicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          }
        }
      },
...
      "CustomerInstanceApplication": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstanceApplicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "application": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Application"
              }
            ]
          },
          "customerInstanceApplicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceApplicationStatus"
              }
            ]
          }
        }
      },
      "CustomerInstance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "customerId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "fullDomain": {
            "type": "string",
            "nullable": true
          },
          "dnsEnvSuffix": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "fullCustomerInstanceName": {
            "type": "string",
            "nullable": true
          },
          "deploymentPhase": {
            "type": "string",
            "nullable": true
          },
          "lifecycleState": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "azureDigitalTwinsInstance": {
            "type": "string",
            "nullable": true
          },
          "azureDataExplorerInstance": {
            "type": "string",
            "nullable": true
          },
          "stampId": {
            "type": "string",
            "format": "guid"
          },
          "customer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Customer"
              }
            ]
          },
          "customerInstanceStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceStatus"
              }
            ]
          },
          "stamp": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Stamp"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          },
          "buildings": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Building"
            }
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "Customer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "salesId": {
            "type": "string",
            "nullable": true
          },
          "customerStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "customerStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerStatus"
              }
            ]
          }
        }
      },
...
      "CustomerInstanceStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "adxDatabaseUrl": {
            "type": "string",
            "nullable": true
          },
          "adtInstanceUrl": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          }
        }
      },
      "Stamp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "subscriptionId": {
            "type": "string",
            "format": "guid"
          },
          "environmentName": {
            "type": "string",
            "nullable": true
          },
          "regionShortName": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "region": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Region"
              }
            ]
          },
          "environment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          }
        }
      },
...
      "Building": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
      "BuildingConnector": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "buildingId": {
            "type": "string",
            "nullable": true
          },
          "connectorId": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "buildingConnectorStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "building": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Building"
              }
            ]
          },
          "connector": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Connector"
              }
            ]
          },
          "buildingConnectorStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/BuildingConnectorStatus"
              }
            ]
          }
        }
      },
      "Connector": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "connectorStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "connectorTypeId": {
            "type": "string",
            "nullable": true
          },
          "connectorStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorStatus"
              }
            ]
          },
          "connectorType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorType"
              }
            ]
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
      "ConnectorStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "ConnectorType": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "nullable": true
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
...
      "CustomerInstanceApplicationStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          }
        }
      }
    }
  }
}